Latest Working Code
Here's the latest, tested and seems to work.  There are parameters in
the config.h file that control various things.  Remember that if you
are using a green board it's a version 1 board.  Red boards are version
2.  You need to change that parameter in both the config.h and the
LED2812.s files.

If you do not have a controller board that has the switches and pots
installed, then uncomment #define FREERUN and it will just cycle
through all of the LED modes.

If you want to turn off the channel scanning, then comment out the
#define PHY_ENABLE_ENERGY_DETECTION.  Note that if you do not have a
controller board, the lantern will still time out to the blue throbbing
mode, but now it shows you which channel it is scanning by lighting it
in white.