Roger S:
Here's the latest code.  I think I figured out what's going on with the
energy detection, though I'm still not convinced that it's working very
well.  The nodes lock onto the controller pretty well though.  When
they are searching, it shows the channels in red and the one they are
looking at in white.  Once they lock on, they start showing the
patterns as before.  So local mode changed from throbbing blue to
flashing red showing the channel scan.  I tried it with two controllers
and the nodes pick up both controllers.  I have not tried two
controllers on the same channel with different encryption keys or PAN
IDs.  I'm a little puzzled that the energy detection on one controller
does not seem to pick up the other controllers radio power, but perhaps
that's just because they are not colliding.

I added a couple of new parameters to the config.h file.  They are for
diagnostics so you can see what's going on inside the chip.